Moved builder inside system
